refactor(activity): fold lookup helpers into existing files
Move local transaction lookup accessors into activity-list-helpers and inline
details JIT selectors in bridgeDetailsUtils and ActivityDetails instead of
adding standalone hook modules.
