+++ app/components/Views/ActivityDetails/ActivityDetails.tsx
@@ -1,5 +1,6 @@
 import React, { useCallback, useEffect, useRef } from 'react';
 import { ScrollView } from 'react-native';
+import { useSelector } from 'react-redux';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { useIsFocused, useNavigation } from '@react-navigation/native';
 import type { AppNavigationProp } from '../../../core/NavigationService/types';
@@ -13,6 +14,11 @@ import {
 } from '@metamask/design-system-react-native';
 import { strings } from '../../../../locales/i18n';
 import { useParams } from '../../../util/navigation/navUtils';
+import { selectTransactionMetadataById } from '../../../selectors/transactionController';
+import type { RootState } from '../../../reducers';
+import {
+  getLocalTransactionMetaId,
+} from '../../../util/activity-adapters';
 // eslint-disable-next-line import-x/no-restricted-paths
 import {
   useBridgeHistoryItemBySrcTxHash,
@@ -31,7 +37,6 @@ import {
 import { ActivityDetailsSelectorsIDs } from './ActivityDetails.testIds';
 import type { ActivityDetailsParams } from './ActivityDetails.types';
 import { useActivityDetailsItem } from './hooks/useActivityDetailsItem';
-import { useLocalTransactionMeta } from './hooks/useActivityTransactionSources';
 import { ActivityDetailsPendingBanner } from './components/ActivityDetailsPendingBanner';
 import { TemplateLoader } from './templates/TemplateLoader';
 
@@ -78,7 +83,12 @@ const ActivityDetails = () => {
   // Pending speed-up / cancel: resolve the live local `TransactionMeta` for the
   // resolved item so the banner reflects current status/gas. Only local EVM
   // items carry a `TransactionMeta`; API / non-EVM items have none (no banner).
-  const pendingTx = useLocalTransactionMeta(item);
+  const pendingMetaId = item ? getLocalTransactionMetaId(item) : undefined;
+  const pendingTx = useSelector((state: RootState) =>
+    pendingMetaId
+      ? selectTransactionMetadataById(state, pendingMetaId)
+      : undefined,
+  );
 
   const {
     speedUpIsOpen,
