+++ app/components/Views/ActivityDetails/templates/bridgeDetailsUtils.ts
@@ -1,6 +1,7 @@
 import type { BridgeHistoryItem } from '@metamask/bridge-status-controller';
 import type { Transaction } from '@metamask/keyring-api';
 import type { TransactionMeta } from '@metamask/transaction-controller';
+import { useSelector } from 'react-redux';
 import {
   type ActivityListItem,
   type TokenAmount,
@@ -9,10 +10,34 @@ import {
   getLocalTransactionInitialMetaId,
   getLocalTransactionMetaId,
 } from '../../../../util/activity-adapters';
+import { selectNonEvmTransactionsForSelectedAccountGroup } from '../../../../selectors/multichain/multichain';
+import { selectTransactionMetadataById } from '../../../../selectors/transactionController';
+import type { RootState } from '../../../../reducers';
 import { findBridgeHistoryItem } from '../../../../util/bridge/findBridgeHistoryItem';
 import { getAssetIdCaipChainId } from '../activityAssetId';
-import { useKeyringTransactionById } from '../hooks/useKeyringTransactionById';
-import { useLocalTransactionMetaById } from '../hooks/useActivityTransactionSources';
+
+function useTransactionMetaById(
+  metaId: string | undefined,
+): TransactionMeta | undefined {
+  return useSelector((state: RootState) =>
+    metaId ? selectTransactionMetadataById(state, metaId) : undefined,
+  );
+}
+
+function useKeyringTransactionById(
+  transactionId: string | undefined,
+): Transaction | undefined {
+  const nonEvmState = useSelector(
+    selectNonEvmTransactionsForSelectedAccountGroup,
+  );
+  if (!transactionId) {
+    return undefined;
+  }
+  const normalizedId = transactionId.toLowerCase();
+  return nonEvmState?.transactions.find(
+    (transaction) => transaction.id.toLowerCase() === normalizedId,
+  );
+}
 
 export function getBridgeDestinationCaipChainId(
   token: TokenAmount | undefined,
@@ -46,7 +71,7 @@ export function useBridgeHistoryItem(
 ) {
   const initialMetaId =
     getLocalTransactionInitialMetaId(item) ?? getLocalTransactionMetaId(item);
-  const initialMeta = useLocalTransactionMetaById(initialMetaId);
+  const initialMeta = useTransactionMetaById(initialMetaId);
   const actionId =
     getLocalTransactionActionId(item) ?? initialMeta?.actionId ?? undefined;
 
@@ -63,7 +88,7 @@ export function useBridgeExplorerSheetTx(
 ): { evmTxMeta?: TransactionMeta; multiChainTx?: Transaction } {
   const initialMetaId =
     getLocalTransactionInitialMetaId(item) ?? getLocalTransactionMetaId(item);
-  const evmTxMeta = useLocalTransactionMetaById(initialMetaId);
+  const evmTxMeta = useTransactionMetaById(initialMetaId);
   const multiChainTx = useKeyringTransactionById(getKeyringTransactionId(item));
   return { evmTxMeta, multiChainTx };
 }
